开发代码分割功能;优化岛屿增删改查功能;修改ws链接功能

+++ src/components/DropButton/index.tsx
@@ -4,6 +4,7 @@ import { useBookState } from '../../store'
 import { useRequest } from 'ahooks'
 import { Popconfirm } from 'antd';
 import { useTranslation } from 'react-i18next';
+import { requestor } from '../../utils/requestor';
 
 interface DropButtonProps extends React.RefAttributes<HTMLElement> {
     keys: Set<string | undefined>;
@@ -11,15 +12,34 @@ interface DropButtonProps extends React.RefAttributes<HTMLElement> {
 
 export function useDropBook() {
     const db_instance = useBookState(state => state.db_instance)
+    const currentIsland = useBookState(state => state.currentIsland)
+    const {t} = useTranslation()
     const { runAsync: drop, loading } = useRequest(async (keys?: string[]) => {
         if (!keys) {
             return
         }
         try {
+            if(!currentIsland) {
+                throw Error(t('请选择岛屿'))
+            }
             await db_instance?.transaction('rw', db_instance.book_items, db_instance.book_blob, async () => {
                 await db_instance.book_items.where('hash').anyOf(keys).delete()
                 await db_instance.book_blob.where('id').anyOf(keys).delete()
             })
+            const res = await Promise.allSettled(keys.map(key => requestor({
+                url: '/island/removeBookToIsland',
+                data: {
+                    bookId: key,
+                    islandId: currentIsland,
+                }
+            })))
+            for (const e of res) {
+                if(e.status === 'fulfilled') {
+                    message.success('删除成功')
+                } else {
+                    message.error("删除失败")
+                }
+            }
         } catch (error) {
             message.error(error instanceof Error ? error.message : error)
         }
